### 2019-12-20 #1971 @whedon
[PRE REVIEW]: DNAtools: Tools for Analysing Forensic Genetic DNA Data
**Submitting author:** @mikldk (<a href="http://orcid.org/0000-0002-0234-0266">Mikkel Meyer Andersen</a>)
**Repository:** <a href="https://github.com/mikldk/DNAtools" target ="_blank">https://github.com/mikldk/DNAtools</a>
**Version:** 0.2-1
