### 2019-12-20 e6b15a8
Paper references
