+++ paper/paper.bib
@@ -31,19 +31,19 @@
 }
 
 @article{tvedebrink2012,
-  title = "Analysis of matches and partial-matches in a Danish STR data set",
+  title = {{Analysis of matches and partial-matches in a Danish STR data set}},
   journal = "Forensic Science International: Genetics",
   volume = "6",
   number = "3",
   pages = "387--392",
   year = "2012",
   issn = "1872-4973",
-  doi = "https://doi.org/10.1016/j.fsigen.2011.08.001",
+  doi = "10.1016/j.fsigen.2011.08.001",
   author = "Torben Tvedebrink and Poul Svante Eriksen and James Michael Curran and Helle Smidt Mogensen and Niels Morling"
 }
 
 @article{tvedebrink2014,
-  title = "On the exact distribution of the numbers of alleles in DNA mixtures",
+  title = {{On the exact distribution of the numbers of alleles in DNA mixtures}},
   journal = "Forensic Science International: Genetics Supplement Series",
   volume = "4",
   number = "1",
@@ -51,22 +51,23 @@
   year = "2013",
   note = "Progress in Forensic Genetics 15",
   issn = "1875-1768",
-  doi = "https://doi.org/10.1016/j.fsigss.2013.10.142",
+  doi = "10.1016/j.fsigss.2013.10.142",
   author = "Torben Tvedebrink"
 }
 
 @article{weir2004,
   author = "BS Weir",
-  title = "Matching and partially-matching DNA profiles",
+  title = {{Matching and partially-matching DNA profiles}},
   journal = "Journal of Forensic Sciences",
   volume = "49",
   number = "5",
-  year = "2004"
+  year = "2004",
+  doi = "10.1520/jfs2003039",
 }
 
 @article{weir2007,
   author = "BS Weir",
-  title = "The rarity of DNA Profiles.",
+  title = {{The rarity of DNA Profiles}},
   journal = "The Annals of Applied Statistics",
   volume = "1",
   number = "2",
