### 2019-12-20 #1971 @mikldk
Editor suggestion: @mgymrek
Reviewer suggestions: @standage @tomsing1
