### 2019-12-20 #1971 @labarba
> Editor suggestion: `@mgymrek`

Please note that `@mgymrek` has stepped down from the editorial board. I'm the associated EiC on rotation this week, and am looking for another handling editor.
