### 2019-12-21 #1971 @csoneson
@tomsing1 - would you be interested in reviewing this submission for JOSS?
