### 2020-01-06 mikldk/DNAtools#1 @tomsing1
Found a few typos during my review for JOSS
Perhaps these suggestions are helpful?

- The following sentence is a bit hard to understand, e.g. what does "which" refer to? I suggest to split it into two shorter sentences to clarify: "With increasing number of STR markers, the random match probabilities are in the order of $10^{-20}$, which some people regard the random match probabilities associated with DNA evidence as just too small or basically unsupportable."Maybe: "With increasing number of STR markers, the probability of a random match between to unrelated samples is small (in the order of $10^{-20}$). Some people regard such small probabilities associated with DNA evidence as just too small or basically unsupportable."
- Typo / clarity: "researchers looking at the efficicy of new multiplexes". That should probably be "efficacy"? Maybe the authors can clarify what "efficacy of new multiplexes" means?
- Typos in the `db_vignette` vignette:
      - "Applying this method to yields" - should probably be "Applying the plot method to the
        summary yields".
      - "distribution is interesting part" - should probably be "distribution is interesting".
      - "Both functions implements" - should be "Both functions implement".
- Typos in the `dna_vignette` vignette:
      - "constituted by a abbritary" should be "constituted by an arbitrary".
- Typos in the `noa_vignette` vignette:
      - "which implies is can be" should be "which implies it can be"
