### 2020-01-06 mikldk/DNAtools#1 @tomsing1
One more small suggestion:

One of the citations is not rendered correctly (I believe you used  a comma to separate two citations, but it should be a semicolon): `[@weir2004,@weir2007]` should be `[@weir2004; @weir2007]`.
