### 2020-01-06 46cd848
Typos mentioned in #1
