+++ paper/paper.md
@@ -29,11 +29,11 @@ bibliography: paper.bib
 
 DNA evidence is the pre-eminent tool in the modern forensic scientists toolbox. It is widely accepted by the public as well as in the scientific and legal communities, and it has been instrumental in determining both the innocence and guilt of individuals involved in the legal process. Despite this widespread acceptance there is unease regarding the statistical measures used to evaluate DNA evidence amongst some of members of all these communities. 
 
-The prevailing technology in forensic genetics is that of capillary electrophoresis [@butler2010], which measures the lengths of so-called *short tandem repeat* regions on the DNA (STR markers). With increasing number of STR markers, the random match probabilities are in the order of $10^{-20}$, which some people regard the random match probabilities associated with DNA evidence as just too small or basically unsupportable. The random match probabilities express the probability that a randomly chosen individual has *a specific* DNA profile, given we know that at least one other (usually a person of interest) has this profile [@DJBwoe2]. 
+The prevailing technology in forensic genetics is that of capillary electrophoresis [@butler2010], which measures the lengths of so-called *short tandem repeat* regions on the DNA (STR markers). With increasing number of STR markers, the random match probabilities are small (in the order of $10^{-20}$). Some people regard such small random match probabilities associated with DNA evidence as just too small or basically unsupportable. The random match probabilities express the probability that a randomly chosen individual has *a specific* DNA profile, given we know that at least one other (usually a person of interest) has this profile [@DJBwoe2]. 
 
-The fact that there exists near matches in a database of DNA profiles have caused some concern in the community and among end-users [@troyer2001] (see also the webpage "[Arizona DNA Database Matches](http://dna-view.com/ArizonaMatch.htm)" by Charles Brenner). However, making all pairwise comparisons and counting their coinciding alleles does not relate to a specific DNA profile [@weir2004,@weir2007]. Hence, one has to use the correct probabilities and also account for the fact the number of comparisons to be made between all pairs of profiles for a database of size $N$ is $N(N+1)/2$ . The `DNAtools` package implements the methodology of @tvedebrink2012 for efficient computations of the expectation and variance of the number of matches.
+The fact that there exists near matches in a database of DNA profiles have caused some concern in the community and among end-users [@troyer2001] (see also the webpage "[Arizona DNA Database Matches](http://dna-view.com/ArizonaMatch.htm)" by Charles Brenner). However, making all pairwise comparisons and counting their coinciding alleles does not relate to a specific DNA profile [@weir2004;@weir2007]. Hence, one has to use the correct probabilities and also account for the fact the number of comparisons to be made between all pairs of profiles for a database of size $N$ is $N(N+1)/2$ . The `DNAtools` package implements the methodology of @tvedebrink2012 for efficient computations of the expectation and variance of the number of matches.
 
-The analysis of mixed DNA traces have proven to be one of the most challenging tasks in forensic genetics. DNA mixtures, as they are referred to, are observed biological traces which are comprised of biological material from two or more individuals. Assessing the number of contributors to a DNA mixture is difficult. One indicator is the number of distinct alleles in the stain -- the more alleles the more contributors. `DNAtools` implements the expression of @tvedebrink2014, where the distribution of the number of distinct alleles can be computed, while accounting for subpopulation effects by the $\theta$-correction [@tvedebrink2014]. Equally, researchers looking at the efficicy of new multiplexes are interested in understanding the probability that a mixture which truly consists of $n$ individuals appears to consist of $n-1$ individuals. This might happen, for example, when a two person mixture shows no more than two alleles per locus at every locus in a multiplex. DNAtools allows the rapid, and exact computation, of such probabilities for any number of individuals.
[…]
