+++ paper/paper.md
@@ -31,9 +31,11 @@ DNA evidence is the pre-eminent tool in the modern forensic scientists toolbox. 
 
 The prevailing technology in forensic genetics is that of capillary electrophoresis [@butler2010], which measures the lengths of so-called *short tandem repeat* regions on the DNA (STR markers). With increasing number of STR markers, the random match probabilities are small (in the order of $10^{-20}$). Some people regard such small random match probabilities associated with DNA evidence as just too small or basically unsupportable. The random match probabilities express the probability that a randomly chosen individual has *a specific* DNA profile, given we know that at least one other (usually a person of interest) has this profile [@DJBwoe2]. 
 
-The fact that there exists near matches in a database of DNA profiles have caused some concern in the community and among end-users [@troyer2001] (see also the webpage "[Arizona DNA Database Matches](http://dna-view.com/ArizonaMatch.htm)" by Charles Brenner). However, making all pairwise comparisons and counting their coinciding alleles does not relate to a specific DNA profile [@weir2004;@weir2007]. Hence, one has to use the correct probabilities and also account for the fact the number of comparisons to be made between all pairs of profiles for a database of size $N$ is $N(N+1)/2$ . The `DNAtools` package implements the methodology of @tvedebrink2012 for efficient computations of the expectation and variance of the number of matches.
+The fact that there exists near matches in a database of DNA profiles have caused some concern in the community and among end-users [@troyer2001] (see also the webpage "[Arizona DNA Database Matches](http://dna-view.com/ArizonaMatch.htm)" by Charles Brenner). However, making all pairwise comparisons and counting their coinciding alleles does not relate to a specific DNA profile [@weir2004;@weir2007]. Hence, one has to use the correct probabilities and also account for the fact the number of comparisons to be made between all pairs of profiles for a database of size $N$ is $N(N+1)/2$ . The `DNAtools` package implements the methodology of @tvedebrink2012 for efficient computations of the expectation and variance of the number of matches. 
+To our knowledge, `DNAtools` is the only software that can perform such computations. 
 
-The analysis of mixed DNA traces have proven to be one of the most challenging tasks in forensic genetics. DNA mixtures, as they are referred to, are observed biological traces which are comprised of biological material from two or more individuals. Assessing the number of contributors to a DNA mixture is difficult. One indicator is the number of distinct alleles in the stain -- the more alleles the more contributors. `DNAtools` implements the expression of @tvedebrink2014, where the distribution of the number of distinct alleles can be computed, while accounting for subpopulation effects by the $\theta$-correction [@tvedebrink2014]. Equally, researchers looking at the efficacy of new multiplexes (in this context this is mainly about the number of and frequencies of alleles in newly included loci) are interested in understanding the probability that a mixture which truly consists of $n$ individuals appears to consist of $n-1$ individuals. This might happen, for example, when a two person mixture shows no more than two alleles per locus at every locus in a multiplex. DNAtools allows the rapid, and exact computation, of such probabilities for any number of individuals.
[…]
