### 2020-01-06 6149470
Update paper.md
Have changed some of the stuff around RMPs
