+++ paper/paper.md
@@ -29,7 +29,11 @@ bibliography: paper.bib
 
 DNA evidence is the pre-eminent tool in the modern forensic scientists toolbox. It is widely accepted by the public as well as in the scientific and legal communities, and it has been instrumental in determining both the innocence and guilt of individuals involved in the legal process. Despite this widespread acceptance there is unease regarding the statistical measures used to evaluate DNA evidence amongst some of members of all these communities. 
 
-The prevailing technology in forensic genetics is that of capillary electrophoresis [@butler2010], which measures the lengths of so-called *short tandem repeat* regions on the DNA (STR markers). With increasing number of STR markers, the random match probabilities are small (in the order of $10^{-20}$). Some people regard such small random match probabilities associated with DNA evidence as just too small or basically unsupportable. The random match probabilities express the probability that a randomly chosen individual has *a specific* DNA profile, given we know that at least one other (usually a person of interest) has this profile [@DJBwoe2]. 
+The prevailing technology in forensic genetics is that of capillary electrophoresis [@butler2010], which measures the lengths of so-called *short tandem repeat* regions on the DNA (STR markers). One quantity of interest is the random match probability. The random match probability is defined as the probability that a randomly chosen individual has *a specific* DNA profile ($G_C$), given we know that at least one other (usually a person of interest, $S$) has this profile [@DJBwoe2].  We write this as
+\[
+\Pr(G_C|S \equiv G_C).
+\]
+Esimates of this probability become very small (in the order of $10^{-20}$) as the number of STR markers increases. Some people regard the small random match probabilities associated with DNA evidence as just too small or basically unsupportable. 
 
 The fact that there exists near matches in a database of DNA profiles have caused some concern in the community and among end-users [@troyer2001] (see also the webpage "[Arizona DNA Database Matches](http://dna-view.com/ArizonaMatch.htm)" by Charles Brenner). However, making all pairwise comparisons and counting their coinciding alleles does not relate to a specific DNA profile [@weir2004;@weir2007]. Hence, one has to use the correct probabilities and also account for the fact the number of comparisons to be made between all pairs of profiles for a database of size $N$ is $N(N+1)/2$ . The `DNAtools` package implements the methodology of @tvedebrink2012 for efficient computations of the expectation and variance of the number of matches. 
 To our knowledge, `DNAtools` is the only software that can perform such computations. 
