### 2020-01-09 5ebdd81
notation: S referred to person, G_S to its genotype. The previous \equiv G_C$ did not make sense
