### 2020-01-09 3edf989
Update paper.md
I have added two new paragraphs and pointed out how the problem arose, and what the package does to address it.
