### 2020-01-14 mikldk/DNAtools#1 @mikldk
Thanks @tomsing1 ! We have fixed the issues you raise in [these commits](https://github.com/mikldk/DNAtools/compare/e6b15a...43883a). I hope you agree?
