+++ paper/paper.md
@@ -21,7 +21,7 @@ affiliations:
    index: 1
  - name: Department of Statistics, University of Auckland, New Zealand
    index: 2
-date: 10 Dec 2019
+date: 14 Jan 2020
 bibliography: paper.bib
 ---
 
@@ -30,9 +30,9 @@ bibliography: paper.bib
 DNA evidence is the pre-eminent tool in the modern forensic scientists toolbox. It is widely accepted by the public as well as in the scientific and legal communities, and it has been instrumental in determining both the innocence and guilt of individuals involved in the legal process. Despite this widespread acceptance there is unease regarding the statistical measures used to evaluate DNA evidence amongst some of members of all these communities. 
 
 The prevailing technology in forensic genetics is that of capillary electrophoresis [@butler2010], which measures the lengths of so-called *short tandem repeat* regions on the DNA (STR markers). One quantity of interest is the random match probability. The random match probability is defined as the probability that a randomly chosen individual has *a specific* DNA profile ($G_C$), given we know that at least one other (usually person $S$ of interest with genotype $G_S$) has this profile [@DJBwoe2].  We write this as
-\[
+$$
 \Pr(G_C \mid G_S \equiv G_C).
-\]
+$$
 Estimates of this probability become very small (in the order of $10^{-20}$) as the number of STR markers increases. Some people regard the small random match probabilities associated with DNA evidence as just too small or basically unsupportable. 
 
 In 2001, Karen Troyer and others [@troyer2001] published a poster reporting the results of a *database matching* exercise. In such an exercise, every profile is compared to every other profile and the number of loci where the two profiles match is recorded. This is a very useful exercise, because, amongst other things it helps laboratories detect potentially erroneous entries in their databases. The Arizona laboratory, where Troyer worked, used the CODIS set of loci, which was a standardized set of 13 STR markers used in many jurisdictions across the US, including federally by the FBI. Troyer et al.'s poster [@troyer2001] reported that a 9 locus match had been found between two apparently unrelated individuals. This was information was seized upon by an enterprising defence lawyer, because at first glance, it seemed to cast doubt on extremely small match probabilities. That is, how could two unrelated individuals, in a database of 65,000 people, have the same (partial) profile, when the probability of this profile was at most 1 in 754 million ($7.54\times 10^{-8})? This issue is nicely summarized by Charles Brenner on his webpage "[Arizona DNA Database Matches](http://dna-view.com/ArizonaMatch.htm)". 
