### 2020-01-14 #1981 @mikldk
@tomsing1 Thanks. We have addressed you points, thanks! Also see [the issue in the repo](https://github.com/mikldk/DNAtools/issues/1).

Regarding

> No mention is made of other software performing a similar task. Perhaps the authors can
comment on the absence of (and hence need for) the package's functionality in the text?

We have added this phrase:

> To our knowledge, `DNAtools` is the only software that can perform such computations.
