### 2020-01-14 mikldk/DNAtools#2 @standage
README buries the lede
The first three lines of the "Getting started" section of the main README are, in my opinion, the content that any potential user needs to see first. Installation instructions, issue trackers, examples, etc., are all good, but first and foremost a user needs to see **What does this software do? Who is it designed for?**

In general I think the documentation for DNAtools is very good, but I think this change would make it even better for newcomers.
