### 2020-01-14 mikldk/DNAtools#3 @standage
Instructions for offline install?
Installing DNAtools on a computer with no internet connection isn't possible with the provided instructions. Only an occasional R user and not an R developer, it took me a while to figure this out.

- download DNAtools as a .tar.gz archive from GitHub, transfer to computer using removable media
- install `devtools` and DNAtools pre-requisites (multicool, Rcpp, RcppParallel, RcppProgress, Rsolnp)
- install DNAtools in R using `devtools::install_local`

It may be helpful for you to include some instructions for offline installation.
