### 2020-01-14 #1981 @standage
The authors describe DNAtools, an R package that implements methods for analysis of short tandem repeat (STR) profiles obtained from forensic DNA traces. DNAtools advertises two primary capabilities. The first focuses on estimating the complexity of DNA mixtures (the number of contributors) by examination of the total number of observed alleles present in a sample. The second focuses on empirical derivation of match probabilities between pairs of entries in a forensic DNA database, including adjustments for multiple testing. As is customary for R packages, these capabilities are demonstrated in a handful of vignettes prepared using a representative data set.

I have added several comments below for the authors' consideration, but in any case I endorse this submission for publication in JOSS.

## Installation

The standard installation instructions should be sufficient for the majority of users. I am likely a bit of a special case, in that the computers I use for data analysis are airgapped and cannot connect to the internet. As a consequence, `devtools::install_gitub("mikldk/DNAtools")` was not an option for me. However, it's not unreasonable to expect that scientists in some forensics labs may also have an airgapped setup for security reasons. So providing install instructions for this use case is potentially useful.

I'm only an occasional R user and I don't write R packages, so figuring this out took some time and effort. It may be trivial for an R developer to do, or there may even be a better way. But with a bit of research and trial-and-error I was able to install DNAtools successfully using the following procedure.

- download DNAtools as a .tar.gz archive from GitHub, transfer to computer using removable media
- install `devtools` and DNAtools pre-requisites (multicool, Rcpp, RcppParallel, RcppProgress, Rsolnp)
- install DNAtools in R using `devtools::install_local`

I'd recommend including some guidelines for offline installation in the instructions.

Related issue: https://github.com/mikldk/DNAtools/issues/3

## Minor grammatical errors

- JOSS paper
    - "forensic scientists toolbox" should be punctuated "forensic scientist's toolbox"
    - "effificy of new multiplexes" should be spelled "efficacy of new multiplexes"
    - "The analysis of mixed DNA traces *have* proven" should be "The analysis of mixed DNA traces *has* proven"
    - "there *exists* near matches" should be "there *exist* near matches"
- dna_vignette
    - "the number of distint alleles present" should be spelled "the number of distinct alleles present"

## Suggested revisions

In the second paragraph of the JOSS paper, the random match probability (RMP) is used before it is defined. I'd suggest revising the second sentence of that paragraph for clarity (possible example below), and then switching the order of the 2nd and 3rd sentences so that the RMP is defined before it is discussed.

> *The number of STR markers assayed by contemporary forensic kits commonly result in RMPs in the order of 10<sup>-20</sup>, a level of confidence considered insufficiently conclusive by some.*

In general I think the DNAtools documentation is good. But the purpose of the software is buried in the GitHub repo's main README. I'd suggest highlighting the "Getting started" section more prominently. Related issue: https://github.com/mikldk/DNAtools/issues/2.
