### 2020-01-14 #1981 @standage
I have not looked at the comments in @tomsing1's review, nor the author responses. It's possible some of the issues I raise have already been addressed. Please feel free to point this out if it's the case.

Unresolved points.
