### 2020-01-14 b08a86d
Related to #2 and #3
