### 2020-01-14 mikldk/DNAtools#2 @mikldk
Good point. Thank you. Please see b08a86df031e13a634373ddb6e10c169fee00cdf. I hope this addresses this issue?
