### 2020-01-14 mikldk/DNAtools#3 @standage
LGTM!
