### 2020-01-14 mikldk/DNAtools#2 @standage
LGTM!
