### 2020-01-14 #1981 @standage
Looks like the references render correctly in the latest proof.
