### 2020-01-14 252d94d
Minor grammatical errors
