### 2020-01-14 #1981 @standage
Latest proof now also indicates the absence of other tools for comparison.
