### 2020-01-14 #1981 @mikldk
I believe that "Installation" was addressed by https://github.com/mikldk/DNAtools/issues/3.

The "Minor grammatical errors" was addressed by https://github.com/mikldk/DNAtools/commit/252d94d5ac97e90b7be31d94175e5e4a1f865e1e.

I am not sure if "Suggested revisions" is still present for the latest proof?
