### 2020-01-14 #1981 @standage
Yep. I think I misunderstood the sentence for which I was suggesting revisions. The latest proof defines the RMP before discussing it, and is much clearer than the original submission. 👍
