### 2020-01-14 #1981 @standage
It looks like all the errors/typos I pointed out have been addressed. However, a new one was introduced: "surpriseing" should be spelled "surprising."

Sorry for the bikeshedding! :-)
