### 2020-01-14 #1981 @mikldk
> It looks like all the errors/typos I pointed out have been addressed. However, a new one was introduced: "surpriseing" should be spelled "surprising."
>
> Sorry for the bikeshedding! :-)

Thanks! Embarrassing. That's not bikeshedding! Thanks. Now fixed in https://github.com/mikldk/DNAtools/commit/00824acb90611e9041510ef3e902bc6e7b00624b.
