### 2020-01-16 #1981 @csoneson
👋 @mikldk - I have gone through the paper and submitted a PR with a couple of minor fixes. Could you check and merge that, and generate a new proof?

Also, could you please edit the Zenodo archive so that the title and authors exactly match those of the paper.
