### 2020-01-16 #1981 @mikldk
@csoneson Thanks! PR now merged and README/site rebuild: https://github.com/mikldk/DNAtools/commit/d25fc87e7760f6fe7ed65b853618a76b1f966b31 .

I generated new release at GitHub and Zenodo. DOI for new release: DOI: [10.5281/zenodo.3609584](https://doi.org/10.5281/zenodo.3609584) / <https://zenodo.org/record/3609584>.

I hope everything is okay, or else let me know! Thanks for all the help and the contribution.
